add query builder and move to utils
