+++ app/libs/common/src/utils/query-builder.ts
@@ -1,0 +1,9 @@
+export function queryBuilder(queryObject: any, querySchema: any) {
+  const query = {};
+  Object.keys(queryObject).forEach((key) => {
+    if (querySchema[key] && queryObject[key]) {
+      query[key] = queryObject[key];
+    }
+  });
+  return query;
+}
